refactor: extract StravaConsole into focused sub-components
- Extract 10 new components from monolithic StravaConsole.jsx (717 -> 297 lines)
- Create centralized formatters.js utility (formatElapsedTime, formatRelativeTime, formatBytes)
- Fix autoScroll state duplication between component and hook
- Extract mutex error message to helper function in useStravaConsole hook
- New components: ConsoleDisabledView, WarningBanner, RunnerStatusBadge,
  ConnectionStateBadge, RunnerOfflineAlert, ErrorAlert, ConsoleHeader,
  CommandSelectionCard, TerminalPanel
