+++ app/utilities/_components/strava-console/CommandHistoryPanel.jsx
@@ -30,26 +30,7 @@ import {
 } from 'react-icons/md';
 import LogManagementDialog from './LogManagementDialog';
 import { ConfirmDialog } from '../../../_components/ui/ConfirmDialog';
-
-/**
- * Format timestamp to relative time
- */
-function formatRelativeTime(timestamp) {
-  const date = new Date(timestamp);
-  const now = new Date();
-  const diffInMinutes = Math.floor((now - date) / 60000);
-
-  if (diffInMinutes < 1) return 'Just now';
-  if (diffInMinutes < 60) return `${diffInMinutes}m ago`;
-
-  const diffInHours = Math.floor(diffInMinutes / 60);
-  if (diffInHours < 24) return `${diffInHours}h ago`;
-
-  const diffInDays = Math.floor(diffInHours / 24);
-  if (diffInDays < 7) return `${diffInDays}d ago`;
-
-  return date.toLocaleDateString();
-}
+import { formatRelativeTime } from './utils/formatters';
 
 /**
  * Get status badge props
